+++ NEWS.md
@@ -41,6 +41,7 @@ Version 0.18b1 - dev
   in 2D and 3D, [docs](https://ezdxf.mozman.at/docs/addons/binpacking.html)
 - NEW: arguments `height` and `rotation`for factory methods `add_text()` and `add_attdef()`
 - NEW: argument `size_inches` in function `ezdxf.addons.drawing.matplotlib.qsave()`
+- NEW: DXF/DWG converter function `ezdxf.addons.odafc.convert()`
 - CHANGE: keyword only argument `dxfattribs` for factory methods `add_text()` and `add_attdef()`
 - CHANGE: `recover` module - recovered integer and float values are logged as severe errors
 - CHANGE: method `Path.all_lines_to_curve3` replaced by function `path.lines_to_curve3()`